refactor : 리프레시토큰 이용한 액세스토큰 재생성메서드 구현을 위해 Long 타입으로 테스트중

+++ src/main/java/com/example/todolist_backend/domain/User.java
@@ -17,8 +17,8 @@ import java.util.List;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // 기본키를 자동으로 생성해주는 어노테이션
-    private int id;
-//    private Long id;
+//    private int id;
+    private Long id;
 
     @Column(nullable = false, unique = true)
     private String account;
